Hackathon - Health"O"Mation - Final Product (#9)
* Home Page

This is our first page of the website. One the user loads in, they will be greeted with this home page. This is our first commit

* css file for the whole project

On this css file we beautify our webpage. This is our first commit

* Home Page Commit 2

* style.css commit 2

Changes:
-Style for about page
-style for contacts page
-@media for responsiveness
-style for cures page

* Info. page

cure for abdominal pain

* About Us page

* Create bronchitis.html

* Contacts Page

* info page

cure for bronchitis

* info page

prevention of covid-19

* Page displaying all the cures

* info page

cure for diarrhea

* info page

cure for flu

* info page

cure for food borne illness

* Contact page working function

allows the user to actually send messages to the people at Health"O"Mation once the code is visible on the internet

* info page

cure for headache

* quiz

the quiz to test the user on his medical know ledge - html file

* info page

cure for mononucleosis

* info page

cure for obesity

* info page

cure for pink eye

* javascript for quiz

* info page

cure for sinusitis

* info page

cure for skin infection

* info page

cure for sore throat

* styling for quiz

* sustainability section html

* images for the code

* Delete abdominalicon.jpg

* Delete aboutimg.jpg

* Delete background.jpg

* Delete banner.jpg

* Delete brainbanner.jpg

* Delete bronchitisicon.jpg

* Delete coldicon.jpg

* Delete covid19icon.jpg

* Delete diarrhea.jpg

* Delete diarrheaicon.jpg

* Delete flu.jpg

* All the images for our project

* Delete images directory

* all the images for project

* cure info 

for mononucleosis

* cure info

for obesity

* cure info

for abdominal pain

* info

for about page

* cure info

for bronchitis

* cure info

for diarrhea

* cure info

for sinusitis

* cure info

for skin infection

* cure info

for sore throat

* style

* sustainability info